requirements.txt
pyqt5
# Main requirements for WreckCheck
flask
pillow
scikit-learn
torch
torchvision
